feat: harden auth flow and stabilize admin data loading
Apply round-7 security and reliability fixes across the admin app, including strict tenant/permission defaults, safer shop switching, guarded async flows, and stale-response protection for paged queries to prevent data corruption and unintended access.

Made-with: Cursor

+++ nexus-oa/src/main/java/com/nexus/oa/application/dto/OaDtos.java
@@ -170,7 +170,7 @@ public final class OaDtos {
         private Integer earlyLeaveDays;    // 早退次数
         private Integer absentDays;        // 旷工天数
         private Integer missingCardDays;   // 缺卡次数
-        private Integer overtimeHours;     // 加班时长（小时）
+        private BigDecimal overtimeHours;     // 加班时长（小时）
     }
 
     // ===================== 请假（新表） =====================
